Formatting and safety check changes
