+++ packages/mutations-ts/src/utils/index.ts
@@ -1,0 +1,6 @@
+import { hasDirectives, getUniqueMutations} from './schemaParsing'
+
+export {
+    hasDirectives,
+    getUniqueMutations
+}
