+++ packages/mutations-ts/src/utils/schemaParsing.ts
@@ -1,5 +1,6 @@
 import { DocumentNode } from 'graphql'
 import { visit } from 'graphql/language/visitor';
+import { makeRepeatedUnique } from './arrayUtils';
 
 export function getDirectiveNames(doc: DocumentNode) {
   const names: string[] = [];
@@ -13,20 +14,20 @@ export function getDirectiveNames(doc: DocumentNode) {
   return names;
 }
 
-export function getMutations(doc: DocumentNode) {
+export function getUniqueMutations(doc: DocumentNode, resolverNames: string[]) {
   let names: string [] = [];
   visit(doc, {
     OperationDefinition(node) {
-      names = node.selectionSet.selections.reduce(function(result: string[], selection){
-        if(selection.kind === "Field"){
+      node.selectionSet.selections.reduce(function(result: string[], selection){
+        if(selection.kind === "Field" && resolverNames.includes(selection.name.value)){
           result.push(selection.name.value)
         }
         return result;
-      }, [])
+      }, names)
     },
   })
 
-  return handleRepeated(names)
+  return makeRepeatedUnique(names)
   
 }
 
@@ -34,15 +35,4 @@ export function hasDirectives(names: string[], doc: DocumentNode) {
   return getDirectiveNames(doc).some(
     (name: string) => names.indexOf(name) > -1,
   );
-}
-
-function handleRepeated(array: string[]){
-  const map: any = {};
-  const count = array.map(function(val) {
-      return map[val] = (typeof map[val] === "undefined") ? 1 : map[val] + 1;
-  });
-
-  return array.map(function(val, index) {
-      return val + (map[val] != 1 ? '_' + count[index] : '');
-  });
 }
